requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
twilio==8.10.0
Pillow==10.0.0
bcrypt==4.0.1
requests==2.31.0
python-jose==3.3.0
gunicorn==21.2.0
pytest==7.4.2
black==23.7.0
flake8==6.1.0
python-magic==0.4.27
geopy==2.4.0
PyJWT==2.8.0
cryptography==41.0.3
python-multipart==0.0.6
validators==0.22.0 